moved timeAgo to nsUtils nsUtils is now in the bundle, mapped to Nightscout namespace cannula age now considers the point of time being visualized in the timeline small prettification on client.js
