+++ lib/plugins/cannulaage.js
@@ -17,14 +17,14 @@ function init() {
     var found = false;
     var treatmentDate = null;
     var message = '';
-
+    
     for (var t in this.env.treatments) {
       if (this.env.treatments.hasOwnProperty(t)) {
         var treatment = this.env.treatments[t];
 
         if (treatment.eventType == "Site Change") {
           treatmentDate = new Date(treatment.created_at);
-          var hours = Math.round(Math.abs(new Date() - treatmentDate) / 36e5);
+          var hours = Math.round(Math.abs(this.env.time - treatmentDate) / 36e5);
 
           if (!found) {
             found = true;
